Models have been siphoned out
