+++ Multiplatform/macOS/Preferences/Preference Panes/Accounts/Account Preferences/Add Account/Add Account Models/AddFeedbinViewModel.swift
@@ -1,0 +1,21 @@
+//
+//  AddFeedbinViewModel.swift
+//  Multiplatform macOS
+//
+//  Created by Stuart Breckenridge on 05/12/2020.
+//  Copyright © 2020 Ranchero Software. All rights reserved.
+//
+
+import SwiftUI
+import Account
+import RSCore
+import RSWeb
+import Secrets
+
+class AddFeedbinViewModel: ObservableObject {
+	@Published var isAuthenticating: Bool = false
+	@Published var accountUpdateError: AccountUpdateErrors = .none
+	@Published var showError: Bool = false
+	@Published var username: String = ""
+	@Published var password: String = ""
+}
