+++ Multiplatform/macOS/Preferences/Preference Panes/Accounts/Account Preferences/Add Account/Add Account Sheets/AddFeedWranglerAccountView.swift
@@ -1,5 +1,5 @@
 //
-//  AddNewsBlurAccountView.swift
+//  AddFeedWranglerAccountView.swift
 //  Multiplatform macOS
 //
 //  Created by Stuart Breckenridge on 03/12/2020.
@@ -12,36 +12,29 @@ import RSCore
 import RSWeb
 import Secrets
 
-fileprivate class AddNewsBlurViewModel: ObservableObject {
-	@Published var isAuthenticating: Bool = false
-	@Published var accountUpdateError: AccountUpdateErrors = .none
-	@Published var showError: Bool = false
-	@Published var username: String = ""
-	@Published var password: String = ""
-}
 
-
-struct AddNewsBlurAccountView: View {
-	
+struct AddFeedWranglerAccountView: View {
+    
 	@Environment (\.presentationMode) var presentationMode
-	@StateObject private var model = AddNewsBlurViewModel()
+	@StateObject private var model = AddFeedWranglerViewModel()
 	
-    var body: some View {
+	var body: some View {
 		VStack {
 			HStack(spacing: 16) {
 				VStack(alignment: .leading) {
-					AccountType.newsBlur.image()
+					AccountType.feedWrangler.image()
+						.resizable()
 						.frame(width: 50, height: 50)
 					Spacer()
 				}
 				VStack(alignment: .leading, spacing: 8) {
-					Text("Sign in to your NewsBlur account.")
+					Text("Sign in to your Feed Wrangler account.")
 						.font(.headline)
 					HStack {
-						Text("Don't have a NewsBlur account?")
+						Text("Don't have a Feed Wrangler account?")
 							.font(.callout)
 						Button(action: {
-							NSWorkspace.shared.open(URL(string: "https://newsblur.com")!)
+							NSWorkspace.shared.open(URL(string: "https://feedwrangler.net/users/new")!)
 						}, label: {
 							Text("Sign up here.").font(.callout)
 						}).buttonStyle(LinkButtonStyle())
@@ -78,7 +71,7 @@ struct AddNewsBlurAccountView: View {
 						}).keyboardShortcut(.cancelAction)
 
 						Button(action: {
-							presentationMode.wrappedValue.dismiss()
+							authenticateFeedWrangler()
 						}, label: {
 							Text("Sign In")
 								.frame(width: 60)
@@ -97,11 +90,14 @@ struct AddNewsBlurAccountView: View {
 		})
     }
 	
-	private func authenticateNewsBlur() {
-		model.isAuthenticating = true
-		let credentials = Credentials(type: .newsBlurBasic, username: model.username, secret: model.password)
+	
+	private func authenticateFeedWrangler() {
 		
-		Account.validateCredentials(type: .newsBlur, credentials: credentials) { result in
+		model.isAuthenticating = true
+		let credentials = Credentials(type: .feedWranglerBasic, username: model.username, secret: model.password)
+		
+		Account.validateCredentials(type: .feedWrangler, credentials: credentials) { result in
+			
 			
 			self.model.isAuthenticating = false
 			
@@ -114,11 +110,11 @@ struct AddNewsBlurAccountView: View {
 					return
 				}
 				
-				let account = AccountManager.shared.createAccount(type: .newsBlur)
+				let account = AccountManager.shared.createAccount(type: .feedWrangler)
 				
 				do {
-					try account.removeCredentials(type: .newsBlurBasic)
-					try account.removeCredentials(type: .newsBlurSessionId)
+					try account.removeCredentials(type: .feedWranglerBasic)
+					try account.removeCredentials(type: .feedWranglerToken)
 					try account.storeCredentials(credentials)
 					try account.storeCredentials(validatedCredentials)
 					account.refreshAll(completion: { result in
@@ -144,8 +140,8 @@ struct AddNewsBlurAccountView: View {
 	}
 }
 
-struct AddNewsBlurAccountView_Previews: PreviewProvider {
+struct AddFeedWranglerAccountView_Previews: PreviewProvider {
     static var previews: some View {
-        AddNewsBlurAccountView()
+        AddFeedWranglerAccountView()
     }
 }
